+++ docs/create.py
@@ -101,6 +101,7 @@ The following packages are required to run fuc:
    pysam
    scipy
    seaborn
+   statsmodels
 
 There are various ways you can install fuc. The recommended way is via conda (`Anaconda <https://www.anaconda.com/>`__):
 
@@ -227,7 +228,7 @@ To merge VCF files:
 
    $ fuc vcf-merge 1.vcf 2.vcf 3.vcf > merged.vcf
 
-To filter a VCF file annotated by Ensemble VEP:
+To filter a VCF file annotated by Ensembl VEP:
 
 .. code-block:: text